Separate StringCodec to separate file
