+++ src/storage/mod.rs
@@ -2,6 +2,7 @@
 mod codec_json;
 #[cfg(feature = "prost")]
 mod codec_prost;
+mod codec_string;
 mod use_storage;
 
 pub use crate::core::StorageType;
@@ -9,4 +10,5 @@ pub use crate::core::StorageType;
 pub use codec_json::*;
 #[cfg(feature = "prost")]
 pub use codec_prost::*;
+pub use codec_string::*;
 pub use use_storage::*;
